backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.10.4",
    "django>=5.2.8",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "google-generativeai>=0.8.6",
    "googletrans==4.0.0rc1",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "redis>=7.1.0",
    "requests>=2.32.5",
]
